nvim check

+++ setup.sh
@@ -19,6 +19,10 @@ install_ripgrep()
 
 install_nvim()
 {
+	if [ -f /usr/local/bin/nvim ]; then
+		echo "Neovim is already installed"
+		return
+	fi
 	curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim.appimage
 	chmod +x nvim.appimage
 	sudo mv nvim.appimage /usr/local/bin/nvim
